packages/backend/src/requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
flask_session==0.5.0
langchain==0.1.0
pydantic==1.10.0
spotipy==2.23.0
Unidecode==1.3.7
openai==0.28.0
gunicorn==21.2.0
Werkzeug==2.2.2
python-dotenv==1.0.0